requirements.txt
django
djangorestframework
lxml
requests
beautifulsoup4
python-dotenv
gunicorn
